complete hm3

+++ controllers/contacts.js
@@ -1,7 +1,6 @@
-const contactsFunctions = require('../models/contacts');
-const { nanoid } = require("nanoid");
-const contactsSchema = require('../schemas/contacts');
+const contactsSchemas = require('../schemas/contacts');
 const Contact = require('../models/contacts');
+const mongoose = require('mongoose');
 console.log(Contact)
 
 async function readAll(req, res, next) {
@@ -15,6 +14,9 @@ async function readAll(req, res, next) {
 
 async function getById (req, res, next) {
   const { contactId } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+      return res.status(400).send({ message: 'Invalid ID' });
+  }
   try {
     const docs = await Contact.findById(contactId).exec();
     if (!docs) { 
@@ -28,6 +30,9 @@ async function getById (req, res, next) {
 
 async function deleteOne (req, res, next) {
   const { contactId } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+      return res.status(400).send({ message: 'Invalid ID' });
+  }
   try {
     const docs = await Contact.findByIdAndDelete(contactId).exec();
     if (!docs) { 
@@ -40,7 +45,7 @@ async function deleteOne (req, res, next) {
 };
 
 async function postNew (req, res, next) {
-  const { error } = contactsSchema.validate(req.body)
+  const { error } = contactsSchemas.contactsSchema.validate(req.body)
   if (error) {
     if (error.details[0].type === "any.required") {
       return res.status(400).send({ message: `missing required ${error.details[0].path[0]} field` });
@@ -48,17 +53,23 @@ async function postNew (req, res, next) {
     return res.status(400).send({ message: error.details[0].message});
   }
   const newContact = {
-    id: nanoid(),
-    ...req.body
+    ...req.body,
+    favorite: false
   }
-  contactsFunctions.addContact(newContact)
-  return res.status(201).send(newContact);
-  
+  try {
+    const docs = await Contact.create(newContact);
+    res.status(201).send(docs);
+  } catch (error) {
+    next(error);
+  }
 };
 
 async function putOne (req, res, next) {
   const { contactId } = req.params;
-  const { error } = contactsSchema.validate(req.body)
+  const { error } = contactsSchemas.contactsSchema.validate(req.body)
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+      return res.status(400).send({ message: 'Invalid ID' });
+  }
 
   if (!Object.keys(req.body).length) { 
     return res.status(400).send({ message: `missing fields` });
@@ -72,11 +83,38 @@ async function putOne (req, res, next) {
   const newContact = {
     ...req.body
   }
-  const updatedContact = await contactsFunctions.updateContact(contactId, newContact)
-  if (!updatedContact) {
+  const doc = await Contact.findByIdAndUpdate(contactId, newContact, { new: true }).exec();
+  if (!doc) {
     return res.status(404).send({ message: 'Not found' })
   }
-  return res.status(200).send(updatedContact);
+  return res.status(200).send(doc);
+};
+
+async function updateStatusContact (req, res, next) {
+  const { contactId } = req.params;
+  const { error } = contactsSchemas.favoriteUpdSchema.validate(req.body)
+  
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+      return res.status(400).send({ message: 'Invalid ID' });
+  }
+
+  if (!Object.keys(req.body).length) { 
+    return res.status(400).send({ message: `missing field favorite` });
+  }
+  if (error) {
+      if (error.details[0].type === "any.required") {
+        return res.status(400).send({ message: `missing required ${error.details[0].path[0]} field` });
+      }
+      return res.status(400).send({ message: error.details[0].message});
+  }
+  const newContact = {
+    ...req.body
+  }
+  const doc = await Contact.findByIdAndUpdate(contactId, newContact, { new: true }).exec();
+  if (!doc) {
+    return res.status(404).send({ message: 'Not found' })
+  }
+  return res.status(200).send(doc);
 };
 
 module.exports = {
@@ -85,4 +123,5 @@ module.exports = {
   deleteOne,
   putOne,
   postNew,
+  updateStatusContact
 }
